fix bug admin detail absen

+++ app/Http/Controllers/Admin/AttendanceTodayController.php
@@ -2,26 +2,32 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
+use App\Http\Requests\Admin\UpdateAbsenTodayRequest;
 use App\Models\Attendance;
 use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Redirect;
-use App\Http\Requests\Admin\UpdateAbsenTodayRequest;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class AttendanceTodayController extends Controller
 {
     public function index()
     {
+        Gate::authorize('adminDashboard', Auth::user());
+
         return view('admin.attendanceToday.index');
     }
 
     public function detail(Attendance $attendance)
     {
+        Gate::authorize('adminDashboard', Auth::user());
+
         return view('admin.attendanceToday.detail', compact('attendance'));
     }
 
     public function update(UpdateAbsenTodayRequest $request, Attendance $attendance)
     {
+        Gate::authorize('adminDashboard', Auth::user());
         $attendance->fill($request->validated());
 
         $attendance->save();
@@ -31,6 +37,7 @@ class AttendanceTodayController extends Controller
 
     public function destroy(Request $request, Attendance $attendance)
     {
+        Gate::authorize('adminDashboard', Auth::user());
         $attendance->delete();
 
         return redirect()->route('admin.absenToday.index');
