+++ app/Http/Controllers/Admin/PositionController.php
@@ -2,21 +2,26 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\Position;
 use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PositionController extends Controller
 {
     public function index()
     {
+        Gate::authorize('adminDashboard', Auth::user());
+
         return view('admin.jabatan.index');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('adminDashboard', Auth::user());
         $validated = $request->validate([
-            'name' => ['min:3', 'required']
+            'name' => ['min:3', 'required'],
         ]);
 
         Position::create($validated);
@@ -26,6 +31,7 @@ class PositionController extends Controller
 
     public function detail(Position $position)
     {
+        Gate::authorize('adminDashboard', Auth::user());
         $users = $position->users()->where('admin', 0)->get();
 
         return view('admin.jabatan.detail', compact('position', 'users'));
@@ -33,6 +39,7 @@ class PositionController extends Controller
 
     public function destroy(Position $position)
     {
+        Gate::authorize('adminDashboard', Auth::user());
         $position->delete();
 
         return back()->with('status', 'jabatan-deleted');
